moved DataModel abc to its own file and updated config_parser to use it as well as setp.py
